fix(vercel): 新增 api/tsconfig.json 提供 NodeNext 解析 + 修 store.ts 类型
根 tsconfig.json 被 .vercelignore 排除,@vercel/node 类型检查 api 入口时
退回默认 moduleResolution(node),不把 '../server/src/app.js' 改写成 .ts,
报 TS2307: Cannot find module。

新增 api/tsconfig.json(NodeNext),@vercel/node 就近命中,正确解析 .js→.ts。
这把整条 server/src import 链拉进类型检查,暴露 store.ts 两处 TS 5.7+
Buffer↔Uint8Array<ArrayBuffer> 泛型错误(npm run build 从不检查 server/),
用 as Uint8Array cast 修复(纯类型,运行时 Buffer 本就是 Uint8Array)。

本地全链路验证通过:vercel-build.sh 编译 → api 类型检查 exit 0 →
运行时加载 app.js OK → server typecheck exit 0。

+++ server/src/store.ts
@@ -36,7 +36,7 @@ function nodeStreamToWeb(stream: fsSync.ReadStream): ReadableStream<Uint8Array> 
   return new ReadableStream<Uint8Array>({
     start(controller) {
       stream.on('data', (chunk) =>
-        controller.enqueue(typeof chunk === 'string' ? new TextEncoder().encode(chunk) : chunk),
+        controller.enqueue(typeof chunk === 'string' ? new TextEncoder().encode(chunk) : (chunk as Uint8Array)),
       );
       stream.on('end', () => controller.close());
       stream.on('error', (e) => controller.error(e));
@@ -58,7 +58,7 @@ export class LocalStore implements ShareStore {
   }
 
   async writeShare(meta: ShareMeta, zip: Buffer): Promise<void> {
-    await fs.writeFile(zipPath(meta.id), zip);
+    await fs.writeFile(zipPath(meta.id), zip as Uint8Array);
     await fs.writeFile(metaPath(meta.id), JSON.stringify(meta, null, 2), 'utf8');
   }
